+++ ChangeLog.md
@@ -5,6 +5,7 @@
 - Add Autolabeler
 - Bump ReleaseDrafter from 6 to 7
 - Fix SQL Migration Statement
+- Run cmake-format
 - Update to pupnp 1.18.3
 
 ### v3.2.0